requirements.txt
WeasyPrint==52.5
git+https://github.com/comwes/mkpdfs-mkdocs-plugin.git@6b692de17642431087a8c5963023f7816c04c991#egg=mkpdfs_mkdocs
ruamel.yaml==0.17.32
mkdocs==1.5.2
mkdocs-glightbox==0.3.4
mkdocs-material==9.1.21